requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
langchain-openai
python-dotenv
fastapi
uvicorn
streamlit